Views, model-migr & Controller for API from Swagger
Currently one can read from swagger and present to the developer what it
will implement. If there is already an API with this name created by me,
then give me a warning for overwriting.
Next step is to add this API, if only there is not another one with same
name!